+++ sim/src/matmulcl.h
@@ -1,6 +1,6 @@
 #ifndef MATMULCL_H
 #define MATMULCL_H
 
-extern bool matmul(int, bool);
+extern bool matmul(size_t, bool);
 
 #endif // MATMULCL_H
